requirements.txt
streamlit
requests
Pillow
beautifulsoup4
wikipedia-api
